Functions

Functions

Parametron.Functions — Module.

The Functions module provides types that represent decision variables and functions of these decision variables that possess a certain structure, such as being affine or quadratic in the decision variables.

Various operators are overloaded to make it easy to construct such functions.

Example

julia> x, y = Variable.(1 : 2)
2-element Array{Parametron.Functions.Variable,1}:
 Parametron.Functions.Variable(1)
 Parametron.Functions.Variable(2)

julia> fun = 3 * (x + y) + 4
3 * x1 + 3 * x2 + 4

julia> typeof(fun)
Parametron.Functions.AffineFunction{Int64}

julia> fun.linear
2-element Array{Parametron.Functions.LinearTerm{Int64},1}:

 3 * x1
 3 * x2

julia> fun.constant
Base.RefValue{Int64}(4)

The Functions module also provides in-place versions of certain common operations, e.g., add!, subtract!, and vecdot!, which may be used to evaluate operations performed on the functions into a preallocated destination without any heap allocations.

source

Types

Parametron.Functions.Variable — Type.

struct Variable

Represents a single decision variable.

source

Parametron.Functions.LinearTerm — Type.

struct LinearTerm{T}

Represents a scalar linear term, i.e. a decision variable scaled by a coefficient.

source

Parametron.Functions.QuadraticTerm — Type.

struct QuadraticTerm{T}

Represents a scalar quadratic term, i.e. the product of two decision variables scaled by a coefficient.

source

Parametron.Functions.AffineFunction — Type.

struct AffineFunction{T}

A scalar affine function represented by a sum of LinearTerms and a constant.

AffineFunction overloads the call operator, which can be used to evalute the function given values for the decision variables. The call operator takes an AbstractDict{Variable, T} collection, which associates values with variables.

Example

julia> x, y = Variable.(1 : 2);

julia> affinefun = 2 * x + 3 * y + 4
2 * x1 + 3 * x2 + 4

julia> vals = Dict(x => 4, y => -3);

julia> affinefun(vals)
3

source

Parametron.Functions.QuadraticFunction — Type.

struct QuadraticFunction{T}

A scalar quadratic function represented by a sum of QuadraticTerms and an AffineFunction.

QuadraticFunction overloads the call operator, which can be used to evalute the function given values for the decision variables. The call operator takes an AbstractDict{Variable, T} collection, which associates values with variables.

Examples

julia> x, y = Variable.(1 : 2);

julia> quadraticfun = 2 * x^2 + 3 * x * y - 2 * y + 4
2 * x1 * x1 + 3 * x1 * x2 + -2 * x2 + 4

julia> vals = Dict(x => 4, y => -3);

julia> quadraticfun(vals)
6

source

Exported functions

Parametron.Functions.canonicalize — Function.

Re-express a term or function in a canonical form.

source

Parametron.Functions.canonicalize — Method.

canonicalize(term)

Re-express the QuadraticTerm term so that the index of term.rowvar is less than or equal to that of term.colvar.

Example

julia> x, y = Variable.(1 : 2);

julia> term = 3 * y * x
3 * x2 * x1

julia> canonicalize(term)
3 * x1 * x2

source

Parametron.Functions.canonicalize — Method.

canonicalize(f)

Return a canonicalized version of f::AffineFunction, namely with linear terms sorted by variable index and with terms corresponding to the same variable combined.

Example

julia> x, y = Variable.(1 : 2);

julia> f = y + x - 2 * y + 3
1 * x2 + 1 * x1 + -2 * x2 + 3

julia> canonicalize(f)
1 * x1 + -1 * x2 + 3

source

Parametron.Functions.canonicalize — Method.

canonicalize(f)

Return a canonicalized version of f::QuadraticFunction. See canonicalize(f::QuadraticTerm) and canonicalize(f::AffineFunction) for more details. Quadratic terms are ordered lexicographically by (term.rowvar, term.colvar).

Example

julia> x, y = Variable.(1 : 2);

julia> f = x * y + y * x + y + y + x - y
1 * x1 * x2 + 1 * x2 * x1 + 1 * x2 + 1 * x2 + 1 * x1 + -1 * x2 + 0

julia> canonicalize(f)
2 * x1 * x2 + 1 * x1 + 1 * x2 + 0

source

Parametron.Functions.canonicalize! — Function.

In-place version of canonicalize.

source

In-place math functions (unexported)

Parametron.Functions.add! — Function.

Add x to f, modifying f.

add!(f, x)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:420.

add!(f, x)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:421.

add!(f, x)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:422.

add!(f, x)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:423.

add!(f, x)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:425.

add!(f, x)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:426.

add!(f, x)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:427.

add!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:429.

source

Parametron.Functions.subtract! — Function.

Subtract x from f, modifying f.

subtract!(f, x)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:442.

subtract!(f, x)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:443.

subtract!(f, x)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:444.

subtract!(f, x)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:446.

subtract!(f, x)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:455.

subtract!(f, x)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:456.

subtract!(f, x)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:457.

subtract!(f, x)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:458.

subtract!(f, x)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:459.

subtract!(f, x)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:461.

subtract!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:470.

source

Parametron.Functions.muladd! — Function.

Multiply x by y and add the result to dest.

muladd!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:484.

muladd!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:492.

muladd!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:495.

muladd!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:503.

muladd!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:506.

muladd!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:514.

muladd!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:517.

source

Parametron.Functions.vecdot! — Function.

Take the dot product of vectors x and y, storing the result in dest.

vecdot!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:675.

vecdot!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:678.

vecdot!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:680.

vecdot!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:682.

vecdot!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:684.

vecdot!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:686.

vecdot!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:688.

vecdot!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:690.

vecdot!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:692.

source

Parametron.Functions.vecadd! — Function.

Add vector x to vector y, storing the result in dest.

vecadd!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:716.

source

Parametron.Functions.vecsubtract! — Function.

Subtract vector y from vector x, storing the result in dest.

vecsubtract!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:716.

source

Parametron.Functions.matvecmul! — Function.

Compute the matrix-vector product A * x, storing the result in y.

matvecmul!(y, A, x)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:741.

matvecmul!(y, A, x)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:766.

matvecmul!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:790.

source

Parametron.Functions.bilinearmul! — Function.

Compute the bilinear form transpose(x) * A * y, storing the result in dest.

bilinearmul!(dest, Q, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:807.

source

Parametron.Functions.scale! — Function.

Scale a vector by a number and store the result in dest.

scale!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:839.

scale!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:850.

scale!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:861.

scale!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:872.

scale!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:880.

scale!(dest, x, y)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:885.

source

Parametron.Functions.vcat! — Function.

Vertically concatenate a number of vectors, storing the result in y.

vcat!(y, args)

defined at /home/travis/build/tkoolen/Parametron.jl/src/functions.jl:951.

source